Rename api example to yoda state example
