+++ src/routes.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import Home from './features/home';
-import ApiCalls from './features/api-calls/containers';
+import YodaStateExample from './features/yoda-state-example/containers';
 
 const Routes = () =>
   <Switch>
     <Route exact path="/" component={Home} />
-    <Route exact path="/api-calls" component={ApiCalls} />
+    <Route exact path="/yoda-state-example" component={YodaStateExample} />
     <Route path="*" component={Home} />
   </Switch>;
